code/web_application/requirements.txt
Werkzeug==2.2.0
Flask==2.2.0
Flask-Session
Flask-Login
Flask-WTF
WTForms-Alchemy
Flask-SQLAlchemy
Flask-Migrate
Flask-APScheduler
python-dotenv
flask-cors
flask-gtts
country-list
pyopenssl
gunicorn

mysql-connector-python
redshift-connector
email_validator
neverbounce-sdk
passlib
phonenumbers
babel
uuid
boto3
requests
numpy==1.23.1
pandas==1.3.5
backports.zoneinfo==0.2.1
bokeh==2.4.3

argparse
pytest





